% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmm-reg.R
\name{cmm_reg}
\alias{cmm_reg}
\alias{cmm_reg_control}
\alias{logLik.cmm_reg}
\alias{AIC.cmm_reg}
\alias{coef.cmm_reg}
\alias{vcov.cmm_reg}
\alias{print.cmm_reg}
\title{Conway Maxwell Multinomial Regression}
\usage{
cmm_reg(y, m, X, W, base = 1, par_init = NULL, control = NULL)

cmm_reg_control(tol = 1e-08, verbose = FALSE, max_iter = 200)

\method{logLik}{cmm_reg}(object, ...)

\method{AIC}{cmm_reg}(object, ..., k = 2)

\method{coef}{cmm_reg}(object, ...)

\method{vcov}{cmm_reg}(object, extended = FALSE, ...)

\method{print}{cmm_reg}(x, ...)
}
\arguments{
\item{y}{TBD}

\item{m}{TBD}

\item{X}{TBD}

\item{W}{TBD}

\item{base}{TBD}

\item{par_init}{TBD}

\item{control}{TBD}
}
\value{

}
\description{
Functions for CMM Regression.
}
\examples{
print("TBD")

}
