% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmm-reg.R
\name{cmm_reg_control}
\alias{cmm_reg_control}
\title{Conway Maxwell Multinomial Regression Control}
\usage{
cmm_reg_control(base = 1, tol = 1e-08, verbose = FALSE, max_iter = 200)
}
\arguments{
\item{base}{in an integer which specifies which category is considered
the baseline. The default is 1.}

\item{tol}{specifies the convergence tolerance \eqn{\epsilon}. The
default is \code{1e-8}.}

\item{verbose}{is a boolean; if \code{TRUE}, print informative
messages during fitting. Default is \code{FALSE}.}

\item{max_iter}{specifies the maximum number of Newton-Raphson
iterations. Default is \code{200}.}
}
\description{
An object with control arguments for CMM regression.
}
