% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{COMMultReg-package}
\alias{COMMultReg-package}
\title{Conway-Maxwell-Multinomial Regression}
\description{
Conway-Maxwell-Multinomial Regression
}
\details{

}
\references{
Pat M. E. Altham and Robin K. S. Hankin (2012). Multivariate generalizations
of the multiplicative binomial distribution: Introducing the MM package.
Journal of Statistical Software 46.

Darcy Steeg Morris, Andrew M. Raim, and Kimberly F. Sellers (2020+).
A Conway-Maxwell-multinomial distribution for Flexible modeling
of clustered categorical data. (Accepted for publication in
Journal of Multivariate Analysis).
Preprint: https://arxiv.org/abs/1911.02131.

Dirk Eddelbuettel (2013) Seamless R and C++ Integration with Rcpp.
Springer, New York. ISBN

Dirk Eddelbuettel, Conrad Sanderson (2014). RcppArmadillo: Accelerating
R with high-performance C++ linear algebra. Computational Statistics and
Data Analysis, Volume 71, March 2014, pages 1054-1063.

M. Galassi et al, GNU Scientific Library Reference Manual, 3rd Edition,
ISBN 0954612078.
}
