% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_cmb_sample}
\alias{d_cmb_sample}
\title{Compute individual density contributions for
\deqn{
X_i \sim \textrm{CMB}(m_i, p_i, \nu_i),
\quad i = 1, \ldots, n.
}}
\usage{
d_cmb_sample(x, m, p, nu, log = FALSE)
}
\arguments{
\item{x}{An \eqn{n}-dimensional vector of outcomes}

\item{m}{An \eqn{n}-dimensional vector \eqn{m_1, \ldots, m_n}}

\item{p}{An \eqn{n}-dimensional vector of probability parameters
\eqn{p_1, \ldots, p_n}}

\item{nu}{An \eqn{n}-dimensional vector of dispersion parameters
\eqn{\nu_1, \ldots, \nu_n}}

\item{log}{\code{TRUE} or \code{FALSE}; if \code{TRUE}, return the
value on the log-scale.}
}
\value{
A vector of density values
\eqn{f(x_1 \mid m_1, p_1, \nu_1), \ldots, f(x_n \mid m_n, p_n, \nu_n),}
which may be on the log-scale and/or unnormalized
according to input arguments. See \link{cmb}.
}
\description{
Compute individual density contributions for
\deqn{
X_i \sim \textrm{CMB}(m_i, p_i, \nu_i),
\quad i = 1, \ldots, n.
}
}
\examples{
set.seed(1234)

n = 20
m = rep(10, n)

x = rnorm(n)
X = model.matrix(~ x)
beta = c(-1, 1)
p = plogis(X \%*\% beta)

w = rnorm(n)
W = model.matrix(~ w)
gamma = c(0.1, -0.1)
nu = X \%*\% gamma

y = numeric(n)
for (i in 1:n) {
    y[i] = r_cmb(1, m[i], p[i], nu[i])
}

d_cmb_sample(y, m, p, nu, log = TRUE)

}
