% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{inv_mlogit}
\alias{inv_mlogit}
\title{logger}
\usage{
inv_mlogit(x, base = 1)
}
\arguments{
\item{x}{a \eqn{p-1} dimensional vector.}

\item{base}{Index of the baseline category.}
}
\value{
a vector in the \eqn{p} dimensional probability simplex.
}
\description{
Inverse multinomial logit transformation.
}
