% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logger}
\alias{logger}
\title{logger}
\usage{
logger(msg, ...)
}
\arguments{
\item{msg}{format string.}

\item{...}{values to be passed into msg.}
}
\description{
Print a message proceeded by a timestamp, with similar arguments as \code{sprintf}.
}
