% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mlogit}
\alias{mlogit}
\title{logger}
\usage{
mlogit(p, base = 1)
}
\arguments{
\item{p}{a vector in the \eqn{p} dimensional probability simplex.}

\item{base}{Index of the baseline category.}
}
\value{
a \eqn{p-1} dimensional vector.
}
\description{
Inverse multinomial logit transformation.
}
