% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{printf}
\alias{printf}
\title{printf}
\usage{
printf(msg, ...)
}
\arguments{
\item{msg}{format string.}

\item{...}{values to be passed into msg.}
}
\description{
Print a message with similar arguments as \code{sprintf}.
}
