% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cmb.R
\name{cmb}
\alias{cmb}
\alias{d_cmb}
\alias{r_cmb}
\alias{p_cmb}
\alias{q_cmb}
\alias{normconst_cmb}
\alias{e_cmb}
\alias{v_cmb}
\title{Conway Maxwell Binomial distribution}
\usage{
d_cmb(x, m, p, nu, log = FALSE, normalize = TRUE)

r_cmb(n, m, p, nu)

p_cmb(x, m, p, nu)

q_cmb(q, m, p, nu)

normconst_cmb(m, p, nu, log = FALSE)

e_cmb(m, p, nu)

v_cmb(m, p, nu)
}
\arguments{
\item{x}{A scalar representing the outcome.}

\item{m}{Number of trials in the CMB cluster.}

\item{p}{Probability parameter}

\item{nu}{Dispersion parameter}

\item{log}{`TRUE` or `FALSE`; if `TRUE`, return
the value on the log-scale.}

\item{normalize}{`TRUE` or `FALSE`; if `FALSE`, do not
compute or apply the normalizing constant to each density value.}

\item{n}{Number of draws to produce.}
}
\value{
The values returned by each function are:
\itemize{
\item `d_cmb`: a number representing the CMB density \eqn{f(x \mid m, p, \nu)}.
\item `r_cmb`: an \eqn{n}-dimensional vector of draws.
\item `normconst_cmb`: a number representing the normalizing constant \eqn{C(m, p, \nu)}.
\item `e_cmb`: a number representing \eqn{\text{E}(X)}.
\item `v_cmb`: a number representing \eqn{\text{Var}(X)}
}
}
\description{
Functions for \eqn{\text{CMB}(m, p, \nu)} distribution.
}
\details{
A random variable
\eqn{X \sim \text{CMB}(m, p, \nu)} has probability mass function
\deqn{
f(x \mid m, p, \nu) = C(m, p, \nu)^{-1} {m \choose x}^\nu p^{x} (1-p)^{m-x},
\quad x \in \{0, 1, \ldots, m\}
}
with
\deqn{
C(m, p, \nu) = \sum_{x = 0}^m
{m \choose x}^\nu p^{x} (1-p)^{m-x}
}
as the normalizing constant.

CMB can be considered a two-dimensional case of CMM. Furthermore, CMB is
useful in drawing from the general CMM distribution; see
Morris, Raim, and Sellers (2020+).
}
\examples{
set.seed(1234)
m = 10
p = 0.7
nu = 0.8

x = r_cmb(100, m, p, nu)
d_cmb(x[1], m, p, nu, log = TRUE)
normconst_cmb(m, p, nu, log = TRUE)
e_cmb(m, p, nu)
v_cmb(m, p, nu)

}
