% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cmb.R
\name{cmb}
\alias{cmb}
\alias{r_cmb}
\alias{d_cmb}
\alias{d_cmb_normconst}
\alias{loglik_cmb}
\alias{dcmb.old}
\alias{rcmb.old}
\alias{dcmb.slow}
\alias{ecmb}
\alias{ecmb_factorial2}
\title{Conway Maxwell Binomial distribution}
\usage{
r_cmb(n, m, p, nu)

d_cmb(x, m, p, nu, take_log = FALSE, normalize = TRUE)

d_cmb_normconst(m, p, nu, take_log = FALSE)

loglik_cmb(y, m, p, nu)

dcmb.old(x, m, p, nu, take_log = FALSE, normalize = TRUE)

rcmb.old(n, m, p, nu)

dcmb.slow(x, m, p, nu, take_log = FALSE, normalize = TRUE)

ecmb(m, p, nu)

ecmb_factorial2(m, p, nu)
}
\arguments{
\item{m}{}

\item{p}{}

\item{nu}{}

\item{x}{}

\item{normalize}{}

\item{phi}{}
}
\value{

}
\description{
Functions for CMB distribution.
}
