% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cmm-poisreg.R, R/cmm.R
\name{cmm}
\alias{cmm}
\alias{r_cmm}
\alias{d_cmm}
\alias{d_cmm_normconst}
\alias{d_cmm_vectorized}
\alias{loglik_cmm}
\alias{gunterize}
\alias{ecmm}
\alias{vcmm}
\title{Conway Maxwell Multinomial Regression}
\usage{
r_cmm(n, m, p, nu, burn, thin, x_init, report_period)

d_cmm(x, m, p, nu, take_log = FALSE, normalize = TRUE)

d_cmm_normconst(m, p, nu, take_log = FALSE)

d_cmm_vectorized(x, m, p, nu, take_log = FALSE, normalize = TRUE)

loglik_cmm(y, m, P_mat, nu)

gunterize(X, all = FALSE)

ecmm(m, p, nu)

vcmm(m, p, nu)
}
\arguments{
\item{m}{}

\item{nu}{}

\item{x}{}

\item{normalize}{}

\item{y}{}

\item{X}{}

\item{W}{}

\item{base}{}

\item{control}{}

\item{lambda}{}

\item{phi}{}

\item{R}{}

\item{report.period}{}

\item{x.init}{}
}
\value{



}
\description{
Functions for CMM Regression.

Functions for CMM distribution.
}
