% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cmm-reg.R, R/cmm.R
\name{cmm}
\alias{cmm}
\alias{gunterize}
\alias{d_cmm}
\alias{d_cmm_normconst}
\alias{d_cmm_vectorized}
\alias{loglik_cmm}
\alias{r_cmm}
\alias{e_cmm}
\alias{v_cmm}
\title{Conway Maxwell Multinomial Regression}
\usage{
gunterize(X, all = FALSE)

d_cmm(x, m, p, nu, take_log = FALSE, normalize = TRUE)

d_cmm_normconst(m, p, nu, take_log = FALSE)

d_cmm_vectorized(x, m, p, nu, take_log = FALSE, normalize = TRUE)

loglik_cmm(y, m, P_mat, nu)

r_cmm(n, m, p, nu, burn = 0, thin = 1, x_init = NULL, report_period = NULL)

e_cmm(m, p, nu)

v_cmm(m, p, nu)
}
\arguments{
\item{X}{}

\item{x}{}

\item{m}{}

\item{nu}{}

\item{normalize}{}

\item{y}{}

\item{x_init}{}

\item{report_period}{}

\item{W}{}

\item{base}{}

\item{control}{}

\item{lambda}{}

\item{phi}{}

\item{R}{}
}
\value{



}
\description{
Functions for CMM Regression.

Functions for CMM distribution.
}
