% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmm-reg.R
\name{cmm_reg}
\alias{cmm_reg}
\alias{cmm_reg_raw}
\alias{cmm_reg_control}
\alias{logLik.cmm_reg}
\alias{AIC.cmm_reg}
\alias{coef.cmm_reg}
\alias{vcov.cmm_reg}
\alias{get_par_names}
\alias{print.cmm_reg}
\title{Conway Maxwell Multinomial Regression}
\usage{
cmm_reg(
  formula_x,
  formula_w = ~1,
  data = NULL,
  beta_init = NULL,
  gamma_init = NULL,
  control = NULL,
  ...
)

cmm_reg_raw(y, X, W, beta_init = NULL, gamma_init = NULL, control = NULL)

cmm_reg_control(base = 1, tol = 1e-08, verbose = FALSE, max_iter = 200)

\method{logLik}{cmm_reg}(object, ...)

\method{AIC}{cmm_reg}(object, ..., k = 2)

\method{coef}{cmm_reg}(object, extended = FALSE, ...)

\method{vcov}{cmm_reg}(object, extended = FALSE, ...)

get_par_names(object, extended = FALSE, ...)

\method{print}{cmm_reg}(x, ...)
}
\arguments{
\item{formula_x}{TBD}

\item{formula_w}{TBD}

\item{control}{TBD}

\item{y}{TBD}

\item{X}{TBD}

\item{W}{TBD}

\item{extended}{TBD}

\item{par_init}{TBD}
}
\value{

}
\description{
Functions for CMM Regression.
}
\details{
Fit the model
\deqn{
\bm{Y}_i \sim \textrm{CMM}_k(m_i, \bm{p}_i, \nu_i),
\quad i = 1, \ldots, n
}
assuming multinomial logit link
\deqn{
\log\left(\frac{p_{i2}}{p_{i1}} \right) = \bm{x}_i^\top \bm{\beta}_1,
\; \ldots, \;
\log\left(\frac{p_{i,k} }{ p_{i1}} \right) = \bm{x}_i^\top \bm{\beta}_{k-1},
}
and identity link
\deqn{
\nu_i = \bm{w}_i^\top \bm{\gamma}.
}
The first category is assumed to be the baseline by default, but this can be
changed to category \code{a} by specifying the \code{base = a} argument.
}
\examples{
print("TBD")

}
