% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_cmb_sample}
\alias{d_cmb_sample}
\title{Density for CMB random sample}
\usage{
d_cmb_sample(x, m, p, nu)
}
\arguments{
\item{x}{An \eqn{n}-dimensional vector of outcomes}

\item{m}{An \eqn{n}-dimensional vector \eqn{m_1, \ldots, m_n}}

\item{p}{An \eqn{n}-dimensional vector of probability parameters
\eqn{p_1, \ldots, p_n}}

\item{nu}{An \eqn{n}-dimensional vector of dispersion parameters
\eqn{\nu_1, \ldots, \nu_n}}

\item{take_log}{\code{TRUE} or \code{FALSE}; if \code{TRUE}, return the
value on the log-scale.}

\item{normalize}{\code{TRUE} or \code{FALSE}; if \code{FALSE}, do not
compute or apply the normalizing constant to each density value.}
}
\value{
A vector of density values
\eqn{f(x_1 \mid m_1, p_1, \nu_1), \ldots, f(x_n \mid m_n, p_n, \nu_n),}
which may be on the log-scale and/or unnormalized
according to input arguments.
}
\description{
Compute individual density contributions for observations in an independent, but not necessarily
identically distributed, CMB random sample.
}
\examples{
stop("TBD")

}
