% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_cmm_sample}
\alias{d_cmm_sample}
\title{Density for CMM random sample}
\usage{
d_cmm_sample(X, m, P, nu, take_log = FALSE, normalize = TRUE)
}
\arguments{
\item{X}{An \eqn{n \times k} matrix of outcomes, where the \eqn{i}th row
\eqn{\bm{x}_i^\top} represents the \eqn{i}th observation.}

\item{m}{An \eqn{n}-dimensional vector \eqn{m_1, \ldots, m_n}}

\item{P}{An \eqn{n \times k} matrix, where the \eqn{i}th row
\eqn{\bm{p}_i^\top} represents the probability parameter for the
\eqn{i}th observation.}

\item{nu}{An \eqn{n}-dimensional vector of dispersion parameters
\eqn{\nu_1, \ldots, \nu_n}}

\item{take_log}{\code{TRUE} or \code{FALSE}; if \code{TRUE}, return the
value on the log-scale.}

\item{normalize}{\code{TRUE} or \code{FALSE}; if \code{FALSE}, do not
compute or apply the normalizing constant to each density value.}
}
\value{
A vector of density values
\eqn{
f(\bm{x}_1^\top \mid m_1, \bm{p}_1^\top, \nu_1),
\ldots,
f(\bm{x}_n^\top \mid m_n, \bm{p}_n^\top, \nu_n),
}
which may be on the log-scale and/or unnormalized
according to input arguments.
}
\description{
Compute individual density contributions for 
\deqn{
\bm{X}_i \sim \textrm{CMM}_k(m_i, \bm{p}_i, \nu_i),
\quad i = 1, \ldots, n.
}
}
\details{
The entire computation for this function is done in C++, and so may be
more efficient than calling \code{d_cmm} in an loop from R.
}
\examples{
stop("TBD")

}
